fix: gale_sem_give_v2 guard (U-1, H-3) + Rust 1.95 clippy refresh
Primary change (U-1, H-3):
  ffi/src/coarse.rs — gale_sem_give_v2 gated the increment with
  count != limit. A caller passing an out-of-contract pair like
  count=u32::MAX, limit=5 passed the guard and triggered u32 overflow
  (debug panic / release wrap to 0). Tighten to count < limit,
  matching gale::sem::give_decide's precondition and P1 invariant
  (0 <= count <= limit).

  Closes loss-scenario LS-1; satisfies CC-1. Identified by the
  2026-04-21 Mythos bug-hunt pass.

Toolchain catch-up (Rust 1.95 clippy):
  src/systick.rs — max_ticks: allow manual_checked_ops with a comment
    explaining why the three-case branch doesn't collapse into
    checked_div. Verus ensures clauses depend on the explicit shape.
  plain/src/systick.rs, plain/systick.rs — regenerated via verus-strip.
  tests/sched_integration.rs — drop trailing comma (unnecessary_trailing_comma).
  tests/net_buf_integration.rs — size: u32 instead of u16 + cast
    (manual_cast_to_larger).
  tests/proptest_sched.rs — collapse if-into-match arm pattern guard.

Local: 1941 tests pass; clippy -D warnings clean; rivet validate
PASS (0 warnings).

+++ tests/net_buf_integration.rs
@@ -438,10 +438,10 @@ fn add_decide_nb4_invariant() {
 fn push_decide_nb5_invariant() {
     let head_offset: u16 = 16;
     let len: u16 = 20;
-    let size: u16 = 64;
+    let size: u32 = 64;
     let bytes: u16 = 8;
     let (new_head, new_len) = push_decide(head_offset, len, bytes).unwrap();
-    assert!(new_head as u32 + new_len as u32 <= size as u32);
+    assert!(u32::from(new_head) + u32::from(new_len) <= size);
 }
 
 /// NB5: pull_decide preserves bounds.
